HSM: Fixup string passing of signing directives
We pass the return value of get_signing_directives as a string - but then double-quote this in the invocation, mixing arguments as one.

This, rightly, causes OpenSSL to complain.

So, let's drop the erroneous double-quotes, and include a line to silence shellcheck in the various sites where we use the result as an argument.