monitor: Fix and simpify completed and failed device detection
We no longer emit -EXITED statusses, but we do emit -ABORTED statuses in the event of a fatal error.

So, rather than setting a flag in a branch and then checking the flag in another branch - simply add the device to the appropriate list if the appropiate status is signalled.