Changes for Yubikey
derivePublicKey() will generate the key for both PEM and PKCS11

get_signing_directives() is removed and ifs are placed where openssl call is made. This is because get_signing_directives was  returning string that was incorrectly processed by the openssl command (single parameter rather than multiple parameters). There is likely a way to fix the get_signing_directives and avoid repeating the code but my bash skills are too limited to do it unfortunately.

This version of the file was tested with Raspberry Pi 4 used as host (running standard PI OS 64 bit image) and it provisioned CM4 module correctly with the private key stored in the Yubikey 5. One thing that is worth noting is to store the key outside of the 9c slot (I have used 9a) to avoid script getting prompted for pin.