state_recording: Fixup concurrent recording
Set the busy timeout to 5s - allowing a stalled write to wait for up to 5s to perform it's commit.

Additionally, force the transaction to start immediately, avoiding deferred transactions.

Longer-term, a DB host will likely be introduced to serialise the recording in a manner we have better control over.